+++ src/datos/Cliente.java
@@ -4,7 +4,7 @@ import java.util.*;
 
 public class Cliente {
 
-	private int idCliente;
+	private long idCliente;
 	private String nombre;
 	private String apellido;
 	private int dni;
@@ -26,11 +26,11 @@ public class Cliente {
 		this.tipoCliente = tipoCliente;
 	}
 
-	public int getIdCliente() {
+	public long getIdCliente() {
 		return idCliente;
 	}
 
-	protected void setIdCliente(int idCliente) {
+	protected void setIdCliente(long idCliente) {
 		this.idCliente = idCliente;
 	}
